+++ deployer/deploy/task/deploy.php
@@ -4,29 +4,44 @@ namespace Deployer;
 
 task('deploy', [
 
+    // Standard deployer task.
+    'deploy:info',
+
     // Read more on https://github.com/sourcebroker/deployer-extended#deploy-check-lock
     'deploy:check_lock',
 
     // Read more on https://github.com/sourcebroker/deployer-extended#deploy-check-composer-install
     'deploy:check_composer_install',
 
-    // Standard deployer deploy:prepare
+    // Read more on https://github.com/sourcebroker/deployer-extended#deploy-check-branch-local
+    'deploy:check_branch_local',
+
+    // Read more on https://github.com/sourcebroker/deployer-extended#deploy-check-branch
+    'deploy:check_branch',
+
+    // Standard deployer task.
     'deploy:prepare',
 
-    // Standard deployer deploy:lock
+    // Standard deployer task.
     'deploy:lock',
 
-    // Standard deployer deploy:release
+    // Standard deployer task.
     'deploy:release',
 
-    // Standard deployer deploy:update_code
+    // Standard deployer task.
     'deploy:update_code',
 
-    // Standard deployer deploy:shared
+    // Standard deployer task.
     'deploy:shared',
 
-    // Standard deployer deploy:vendors
+    // Standard deployer task.
+    'deploy:writable',
+
+    // Standard deployer task.
     'deploy:vendors',
+
+    // Standard deployer task.
+    'deploy:clear_paths',
 
     // Special for symfony4
     'deploy:database:update',
@@ -37,26 +52,45 @@ task('deploy', [
     // Special for symfony4
     'deploy:cache:warmup',
 
-    // Clear php cli cache.
-    // Read more on https://github.com/sourcebroker/deployer-extended#php-clear-cache-cli
-    'php:clear_cache_cli',
+    // Create database backup, compress and copy to database store.
+    // Read more on https://github.com/sourcebroker/deployer-extended-database#db-backup
+    'db:backup',
 
-    // Standard deployer deploy:writable
-    'deploy:writable',
+    // Start buffering http requests. No frontend access possible from now.
+    // Read more on https://github.com/sourcebroker/deployer-extended#buffer-start
+    'buffer:start',
 
-    // Standard deployer deploy:clear_paths
-    'deploy:clear_paths',
+    // Truncate caching tables, all cf_* tables
+    // Read more on https://github.com/sourcebroker/deployer-extended-database#db-truncate
+    'db:truncate',
 
-    // Standard deployers symlink (symlink release/x/ to current/)
+    // Standard deployer task.
     'deploy:symlink',
 
-    // Standard deployer deploy:unlock
+    // Clear php cli cache.
+    // Read more on https://github.com/sourcebroker/deployer-extended#cache-clear-php-cli
+    'cache:clear_php_cli',
+
+    // Clear frontend http cache.
+    // Read more on https://github.com/sourcebroker/deployer-extended#cache-clear-php-http
+    'cache:clear_php_http',
+
+    // Frontend access possible again from now
+    // Read more on https://github.com/sourcebroker/deployer-extended#buffer-stop
+    'buffer:stop',
+
+    // Standard deployer task.
     'deploy:unlock',
 
-    // Standard deployer cleanup.
+    // Standard deployer task.
     'cleanup',
+
+    // Read more on https://github.com/sourcebroker/deployer-extended#deploy-extend-log
+    'deploy:extend_log',
+
+    // Standard deployer task.
+    'success',
 
 ])->desc('Deploy your Symfony4 project');
 
-after('deploy', 'success');
-
+after('deploy:failed', 'deploy:unlock');